100 Line ET parser using TDOP
This is a 100 line parser with AST and parser capable of matching arbitrary binary operators with varying precedence rules. Adding an additional infix, grouping, and mixed operator type should take about 10 lines each.  That allows matching a nearly abritrary expression language with appropriate binding rules. This seems like an excellent addition to recursive descent parsing for statement classes, which are (typically) more constained in their grammar productions. Using this approach, a grammar can be expressed using RD for statements and switching to TDOP for expression handling, resulting in a compact source code for the language's Front End.